Moves the article endpoints to be in the same module because they will share things
